+++ backend/src/vehiculos/vehiculos.module.ts
@@ -1,9 +1,12 @@
 import { Module } from '@nestjs/common';
+import { PrismaModule } from '../prisma/prisma.module';
 import { VehiculosController } from './vehiculos.controller';
 import { VehiculosService } from './vehiculos.service';
 
 @Module({
+  imports: [PrismaModule],
   controllers: [VehiculosController],
-  providers: [VehiculosService]
+  providers: [VehiculosService],
+  exports: [VehiculosService]
 })
 export class VehiculosModule {}